Toon actuele redenen waarom warmtepompen uit staan (#729)
* Add 'Waarom staat mijn warmtepomp uit?' section to control replay view

This adds a new section before 'Wat doet het systeem daarna?' that shows
all blocking parameters when both heat pumps (HP1 and HP2) are off. The
section displays bullet points with human-readable explanations of why
the heat pump is not running, including:

- Startup inhibit with remaining time
- Minimum rest time
- No candidate available
- Candidate in rest/defrost
- Flow preflow/too low
- Sensor fallback
- Soft guard
- Restart wait
- Frost protection
- Cooling-specific blocking reasons
- Sticky protection
- Generic 'no heat demand' fallback

Each reason is shown as a bullet point in a list, making it clear to users
exactly what needs to change for the heat pump to start running again.

Generated on branch: feature/detailed-stop-info


* Add low-load blocking reasons with stookgrens (heating threshold) details

Extends the 'Waarom staat mijn warmtepomp uit?' section to include:

- Low-load protection: Shows when the heat demand is below the heating
  threshold (stookgrens) with actual wattage values for the off and on
  thresholds. This explains why the pump doesn't start with low heat demand
  in Power House mode.

- CM2 re-entry block: Shows when the system is temporarily blocking re-entry
  to CM2 mode after a low-load stop, including the remaining time.

These additions help users understand the low-load protection mechanism
that prevents short cycling when heat demand is minimal.

Generated on branch: feature/detailed-stop-info


* Update web-budgets.mjs budget

* Style 'Waarom staat mijn warmtepomp uit?' to match 'Wat doet het systeem daarna?'

- Changed the CSS class from oq-working-now-blocking to oq-working-now-next
  to match the styling of 'Wat doet het systeem daarna?'
- Fixed the blocking reasons function to check ALL blocking conditions, not just
  the primaryReason. Now it collects all reasons that block a start:
  - Removed the condition that only checked if primaryReason === 'keep_current'
  - Low-load protection, CM2 re-entry block, and other checks now apply regardless
    of primaryReason, as long as heat pumps are off
  - This ensures all blocking reasons are shown, not just one

Generated on branch: feature/detailed-stop-info

* More detailed info

* Fix compile error from duplicate include and compile warning

* layout

---------

+++ openquatt/includes/control/oq_boiler_runtime.h
@@ -70,7 +70,8 @@ class Runtime {
     id(oq_boiler_rearm_required) = true;
     id(oq_boiler_rearm_after_command_ms) = now_ms;
     withdraw_output_(now_ms);
-    id(oq_boiler_block_reason_code) = available ? oq_boiler::BLOCK_AWAITING_FRESH_COMMAND : unavailable_reason;
+    id(oq_boiler_block_reason_code) =
+        available ? static_cast<uint8_t>(oq_boiler::BLOCK_AWAITING_FRESH_COMMAND) : unavailable_reason;
   }
 
   bool command_is_fresh(uint32_t maximum_age_ms) const {
